Refactor table view controllers to inherit from SwipeTableViewController
