+++ Tasket/CategoriesViewController.swift
@@ -8,9 +8,8 @@
 
 import UIKit
 import RealmSwift
-import SwipeCellKit
 
-class CategoriesViewController: UITableViewController {
+class CategoriesViewController: SwipeTableViewController {
     
     // MARK: - Instance Variables
     let realm = try! Realm()
@@ -21,7 +20,6 @@ class CategoriesViewController: UITableViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         loadData()
-        configureTableView()
     }
 
     // MARK: - Table View Data Source Methods
@@ -31,8 +29,7 @@ class CategoriesViewController: UITableViewController {
     }
 
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "categoryItemCell", for: indexPath) as! SwipeTableViewCell
-        cell.delegate = self
+        let cell = super.tableView(tableView, cellForRowAt: indexPath)
         cell.textLabel?.text = categories?[indexPath.row].name ?? "No Categories Added Yet"
 
         return cell
@@ -53,12 +50,6 @@ class CategoriesViewController: UITableViewController {
         }
     }
     
-    // MARK: - UI Manipulation Methods
-    
-    func configureTableView() {
-        tableView.rowHeight = 80.0
-    }
-    
     // MARK: - Data Manipulation Methods
     
     func save(category: Category) {
@@ -76,6 +67,22 @@ class CategoriesViewController: UITableViewController {
     func loadData() {
         categories = realm.objects(Category.self)
         tableView.reloadData()
+    }
+    
+    override func updateModel(at indexPath: IndexPath) {
+        guard let category = categories?[indexPath.row] else { return }
+        delete(category: category)
+    }
+    
+    func delete(category: Category) {
+        do {
+            try realm.write {
+                realm.delete(category.todoItems)
+                realm.delete(category)
+            }
+        } catch {
+            print("Error deleting category from Realm \(error)")
+        }
     }
     
     // MARK: - IBAction Methods
